// src/enum/filter-model-item-operation.enum.ts
export enum FilterModelItemOperationEnum {
    CONTAINT = 'contains',
    NOTCONTAINT = 'notContains',
    EQUALS = 'equals',
    NOTEQUAL = 'notEqual',
    GREATE = 'greate',
    GREATEEQUALS = 'greateEquals',
    LESS = 'less',
    LESSEQUALS = 'lessEquals',
    STARTSWITH = 'startsWith',
    ENDSWITH = 'endsWith',
    BETWEEN = 'between',
    ISNULL = 'isNull',
    ISNOTNULL = 'isNotNull',
    ISANYOF = 'isAnyOf',
    ISNOTANYOF = 'isNotAnyOf',
    EQUALOPERATOR = '=',
    LESSOPERATOR = '<',
    LESSEQUALOPERATOR = '<=',
    GREATEOPERATOR = '>',
    GREATEEQUALOPERATOR = '>=',
    ISEMPTY = 'isEmpty',
    ISNOTEMPTY = 'isNotEmpty',
  }
  
  export const operatorMappings = {
    [FilterModelItemOperationEnum.CONTAINT]: 'like',
    [FilterModelItemOperationEnum.NOTCONTAINT]: 'not like',
    [FilterModelItemOperationEnum.EQUALS]: '=',
    [FilterModelItemOperationEnum.NOTEQUAL]: '!=',
    [FilterModelItemOperationEnum.GREATE]: '>',
    [FilterModelItemOperationEnum.GREATEEQUALS]: '>=',
    [FilterModelItemOperationEnum.LESS]: '<',
    [FilterModelItemOperationEnum.LESSEQUALS]: '<=',
    [FilterModelItemOperationEnum.STARTSWITH]: 'like',
    [FilterModelItemOperationEnum.ENDSWITH]: 'like',
    [FilterModelItemOperationEnum.BETWEEN]: 'between',
    [FilterModelItemOperationEnum.ISNULL]: 'is null',
    [FilterModelItemOperationEnum.ISNOTNULL]: 'is not null',
    [FilterModelItemOperationEnum.ISANYOF]: 'in',
    [FilterModelItemOperationEnum.ISNOTANYOF]: 'not in',
    [FilterModelItemOperationEnum.EQUALOPERATOR]: '=',
    [FilterModelItemOperationEnum.LESSOPERATOR]: '<',
    [FilterModelItemOperationEnum.LESSEQUALOPERATOR]: '<=',
    [FilterModelItemOperationEnum.GREATEOPERATOR]: '>',
    [FilterModelItemOperationEnum.GREATEEQUALOPERATOR]: '>=',
    [FilterModelItemOperationEnum.ISEMPTY]: 'is null',
    [FilterModelItemOperationEnum.ISNOTEMPTY]: 'is not null',
  };
  